uint32_t -> uint64_t for Board::getMicros()

+++ boards/alienflightf3/alienflightf3.hpp
@@ -105,9 +105,9 @@ class AlienflightF3 : public Board {
         delay(msec);
     }
 
-    virtual uint32_t getMicros() override
+    virtual uint64_t getMicros() override
     {
-        return micros();
+        return (uint64_t)micros();
     }
 
     virtual void ledSet(uint8_t id, bool is_on, float max_brightness) override
