+++ boards/beef/beef.hpp
@@ -96,9 +96,9 @@ class Beef : public Board {
         delay(msec);
     }
 
-    virtual uint32_t getMicros() override
+    virtual uint64_t getMicros() override
     {
-        return micros();
+        return (uint64_t)micros();
     }
 
     virtual void ledSet(uint8_t id, bool is_on, float max_brightness) override
